move apollo provider to index.js to allow queries in App
